adding abstract attribute & disallowing some function calls when abstract

+++ pynamodb/models.py
@@ -9,7 +9,7 @@ import logging
 import warnings
 
 from six import add_metaclass
-from pynamodb.exceptions import DoesNotExist, TableDoesNotExist, TableError
+from pynamodb.exceptions import NotAllowedWhenAbstract, DoesNotExist, TableDoesNotExist, TableError
 from pynamodb.attributes import Attribute, AttributeContainer, AttributeContainerMeta, MapAttribute, ListAttribute
 from pynamodb.connection.base import MetaTable
 from pynamodb.connection.table import TableConnection
@@ -184,6 +184,11 @@ class MetaModel(AttributeContainerMeta):
                         setattr(attr_obj, 'base_backoff_ms', get_settings_value('base_backoff_ms'))
                     if not hasattr(attr_obj, 'max_retry_attempts'):
                         setattr(attr_obj, 'max_retry_attempts', get_settings_value('max_retry_attempts'))
+                    if not hasattr(attr_obj, '_abstract_'):
+                        setattr(attr_obj, '_abstract_', False)
+                    if hasattr(attr_obj, 'table_name'):
+                        if getattr(attr_obj, '_abstract_', False):
+                            raise cls.NotAllowedWhenAbstract("Abstract model can't have a table name")
                 elif issubclass(attr_obj.__class__, (Index, )):
                     attr_obj.Meta.model = cls
                     if not hasattr(attr_obj.Meta, "index_name"):
@@ -192,7 +197,7 @@ class MetaModel(AttributeContainerMeta):
                     if attr_obj.attr_name is None:
                         attr_obj.attr_name = attr_name
 
-            if META_CLASS_NAME not in attrs:
+            if not hasattr(cls, META_CLASS_NAME):
                 setattr(cls, META_CLASS_NAME, DefaultMeta)
 
             # create a custom Model.DoesNotExist derived from pynamodb.exceptions.DoesNotExist,
@@ -203,6 +208,11 @@ class MetaModel(AttributeContainerMeta):
                     exception_attrs['__qualname__'] = '{}.{}'.format(cls.__qualname__, 'DoesNotExist')
                 cls.DoesNotExist = type('DoesNotExist', (DoesNotExist, ), exception_attrs)
 
+            if 'NotAllowedWhenAbstract' not in attrs:
+                exception_attrs = {'__module__': attrs.get('__module__')}
+                if hasattr(cls, '__qualname__'):  # On Python 3, Model.NotAllowedWhenAbstract
+                    exception_attrs['__qualname__'] = '{}.{}'.format(cls.__qualname__, 'NotAllowedWhenAbstract')
+                cls.NotAllowedWhenAbstract = type('NotAllowedWhenAbstract', (NotAllowedWhenAbstract, ), exception_attrs)
 
 @add_metaclass(MetaModel)
 class Model(AttributeContainer):
@@ -227,6 +237,10 @@ class Model(AttributeContainer):
         :param range_key: Only required if the table has a range key attribute.
         :param attrs: A dictionary of attributes to set on this object.
         """
+        # Abstract models aren't allowed to have instance
+        if self.Meta._abstract_:
+            raise self.NotAllowedWhenAbstract()
+
         if hash_key is not None:
             attributes[self._dynamo_to_python_attr(self._get_meta_data().hash_keyname)] = hash_key
         if range_key is not None:
@@ -237,6 +251,11 @@ class Model(AttributeContainer):
                 )
             attributes[self._dynamo_to_python_attr(range_keyname)] = range_key
         super(Model, self).__init__(**attributes)
+
+    @classmethod
+    def dummy(cls):
+        if cls.Meta._abstract_:
+            raise cls.NotAllowedWhenAbstract()
 
     @classmethod
     def has_map_or_list_attributes(cls):
@@ -258,6 +277,8 @@ class Model(AttributeContainer):
         :param items: Should be a list of hash keys to retrieve, or a list of
             tuples if range keys are used.
         """
+        if cls.Meta._abstract_:
+            raise cls.NotAllowedWhenAbstract()
         items = list(items)
         hash_keyname = cls._get_meta_data().hash_keyname
         range_keyname = cls._get_meta_data().range_keyname
@@ -313,9 +334,15 @@ class Model(AttributeContainer):
                             passed here, changes automatically commit on context exit
                             (whether successful or not).
         """
+        if cls.Meta._abstract_:
+            raise cls.NotAllowedWhenAbstract()
         return BatchWrite(cls, auto_commit=auto_commit)
 
     def __repr__(self):
+        # Sanity check, abstract models aren't allowed to have instance
+        if self.__class__.Meta._abstract_:
+            raise self.__class__.NotAllowedWhenAbstract()
+
         if self.Meta.table_name:
             serialized = self._serialize(null_check=False)
             if self._get_meta_data().range_keyname:
@@ -328,6 +355,10 @@ class Model(AttributeContainer):
         """
         Deletes this object from dynamodb
         """
+        # Sanity check, abstract models aren't allowed to have instance
+        if self.__class__.Meta._abstract_:
+            raise self.__class__.NotAllowedWhenAbstract()
+
         self._conditional_operator_check(conditional_operator)
         args, kwargs = self._get_save_args(attributes=False, null_check=False)
         if len(expected_values):
@@ -348,6 +379,10 @@ class Model(AttributeContainer):
             See: http://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_UpdateItem.html#DDB-UpdateItem-request-AttributeUpdate
         """
         warnings.warn("`Model.update_item` is deprecated in favour of `Model.update` now")
+
+        # Sanity check, abstract models aren't allowed to have instance
+        if self.__class__.Meta._abstract_:
+            raise self.__class__.NotAllowedWhenAbstract()
 
         self._conditional_operator_check(conditional_operator)
         args, save_kwargs = self._get_save_args(null_check=False)
@@ -398,6 +433,10 @@ class Model(AttributeContainer):
                                 next_attr: {'value': True, 'action': 'PUT'},
                             }
         """
+        # Sanity check, abstract models aren't allowed to have instance
+        if self.__class__.Meta._abstract_:
+            raise self.__class__.NotAllowedWhenAbstract()
+
         if attributes is not None and not isinstance(attributes, dict):
             raise TypeError("the value of `attributes` is expected to be a dictionary")
         if actions is not None and not isinstance(actions, list):
@@ -444,6 +483,10 @@ class Model(AttributeContainer):
         """
         Save this object to dynamodb
         """
+        # Sanity check, abstract models aren't allowed to have instance
+        if self.__class__.Meta._abstract_:
+            raise self.__class__.NotAllowedWhenAbstract()
+
         self._conditional_operator_check(conditional_operator)
         args, kwargs = self._get_save_args()
         if len(expected_values):
@@ -458,6 +501,10 @@ class Model(AttributeContainer):
 
         :param consistent_read: If True, then a consistent read is performed.
         """
+        # Sanity check, abstract models aren't allowed to have instance
+        if self.__class__.Meta._abstract_:
+            raise self.__class__.NotAllowedWhenAbstract()
+
         args, kwargs = self._get_save_args(attributes=False)
         kwargs.setdefault('consistent_read', consistent_read)
         attrs = self._get_connection().get_item(*args, **kwargs)
@@ -478,6 +525,8 @@ class Model(AttributeContainer):
         :param hash_key: The hash key of the desired item
         :param range_key: The range key of the desired item, only used when appropriate.
         """
+        if cls.Meta._abstract_:
+            raise cls.NotAllowedWhenAbstract()
         hash_key, range_key = cls._serialize_keys(hash_key, range_key)
         data = cls._get_connection().get_item(
             hash_key,
@@ -499,6 +548,8 @@ class Model(AttributeContainer):
 
         :param data: A serialized DynamoDB object
         """
+        if cls.Meta._abstract_:
+            raise cls.NotAllowedWhenAbstract()
         mutable_data = copy.copy(data)
         if mutable_data is None:
             raise ValueError("Received no mutable_data to construct object")
@@ -543,6 +594,8 @@ class Model(AttributeContainer):
         :param index_name: If set, then this index is used
         :param filters: A dictionary of filters to be used in the query. Requires a hash_key to be passed.
         """
+        if cls.Meta._abstract_:
+            raise cls.NotAllowedWhenAbstract()
         if hash_key is None:
             if filters:
                 raise ValueError('A hash_key must be given to use filters')
@@ -624,6 +677,8 @@ class Model(AttributeContainer):
         :param page_size: Page size of the query to DynamoDB
         :param filters: A dictionary of filters to be used in the query
         """
+        if cls.Meta._abstract_:
+            raise cls.NotAllowedWhenAbstract()
         cls._conditional_operator_check(conditional_operator)
         cls._get_indexes()
         if index_name:
@@ -717,6 +772,8 @@ class Model(AttributeContainer):
         :param consistent_read: If True, a consistent read is performed
         """
 
+        if cls.Meta._abstract_:
+            raise cls.NotAllowedWhenAbstract()
         cls._conditional_operator_check(conditional_operator)
         key_filter, scan_filter = cls._build_filters(
             SCAN_OPERATOR_MAP,
@@ -773,6 +830,8 @@ class Model(AttributeContainer):
         :param filters: A list of item filters
         :param consistent_read: If True, a consistent read is performed
         """
+        if cls.Meta._abstract_:
+            raise cls.NotAllowedWhenAbstract()
         cls._conditional_operator_check(conditional_operator)
         key_filter, scan_filter = cls._build_filters(
             SCAN_OPERATOR_MAP,
@@ -811,6 +870,8 @@ class Model(AttributeContainer):
         """
         Returns True if this table exists, False otherwise
         """
+        if cls.Meta._abstract_:
+            raise cls.NotAllowedWhenAbstract()
         try:
             cls._get_connection().describe_table()
             return True
@@ -822,6 +883,8 @@ class Model(AttributeContainer):
         """
         Delete the table for this model
         """
+        if cls.Meta._abstract_:
+            raise cls.NotAllowedWhenAbstract()
         return cls._get_connection().delete_table()
 
     @classmethod
@@ -829,6 +892,8 @@ class Model(AttributeContainer):
         """
         Returns the result of a DescribeTable operation on this model's table
         """
+        if cls.Meta._abstract_:
+            raise cls.NotAllowedWhenAbstract()
         return cls._get_connection().describe_table()
 
     @classmethod
@@ -840,6 +905,8 @@ class Model(AttributeContainer):
         :param read_capacity_units: Sets the read capacity units for this table
         :param write_capacity_units: Sets the write capacity units for this table
         """
+        if cls.Meta._abstract_:
+            raise cls.NotAllowedWhenAbstract()
         if not cls.exists():
             schema = cls._get_schema()
             if hasattr(cls.Meta, pythonic(READ_CAPACITY_UNITS)):
@@ -885,6 +952,8 @@ class Model(AttributeContainer):
         """
         Returns a JSON representation of this model's table
         """
+        if cls.Meta._abstract_:
+            raise cls.NotAllowedWhenAbstract()
         return json.dumps([item._get_json() for item in cls.scan()])
 
     @classmethod
@@ -892,11 +961,15 @@ class Model(AttributeContainer):
         """
         Writes the contents of this model's table as JSON to the given filename
         """
+        if cls.Meta._abstract_:
+            raise cls.NotAllowedWhenAbstract()
         with open(filename, 'w') as out:
             out.write(cls.dumps())
 
     @classmethod
     def loads(cls, data):
+        if cls.Meta._abstract_:
+            raise cls.NotAllowedWhenAbstract()
         content = json.loads(data)
         with cls.batch_write() as batch:
             for item_data in content:
@@ -905,6 +978,8 @@ class Model(AttributeContainer):
 
     @classmethod
     def load(cls, filename):
+        if cls.Meta._abstract_:
+            raise cls.NotAllowedWhenAbstract()
         with open(filename, 'r') as inf:
             cls.loads(inf.read())
 
@@ -1158,6 +1233,9 @@ class Model(AttributeContainer):
         """
         Returns a Python object suitable for serialization
         """
+        # Sanity check, abstract models aren't allowed to have instance
+        if self.Meta._abstract_:
+            raise NotAllowedWhenAbstract()
         kwargs = {}
         serialized = self._serialize(null_check=False)
         hash_key = serialized.get(HASH)
@@ -1176,6 +1254,9 @@ class Model(AttributeContainer):
         :param attributes: If True, then attributes are included.
         :param null_check: If True, then attributes are checked for null.
         """
+        # Sanity check, abstract models aren't allowed to have instance
+        if self.Meta._abstract_:
+            raise NotAllowedWhenAbstract()
         kwargs = {}
         serialized = self._serialize(null_check=null_check)
         hash_key = serialized.get(HASH)
